backend/requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
chromadb>=0.4.0,<0.7.0
uvicorn